requirements.txt
python-dateutil==2.8.1
hypothesis==5.4.1
numpy==1.18.1
pymemcache==3.0.0
guppy3==3.0.9
pandas==1.0.1
